refactor: tracking chain id
