+++ packages/maskbook/src/plugins/Trader/messages.ts
@@ -1,9 +1,9 @@
+import type { WebExtensionMessage } from '@dimensiondev/holoflows-kit'
 import type { TagType, DataProvider, TradeProvider } from './types'
 import { createPluginMessage } from '../utils/createPluginMessage'
 import { PLUGIN_IDENTIFIER } from './constants'
 import { createPluginRPC } from '../utils/createPluginRPC'
 import type { TraderProps } from './UI/trader/Trader'
-import type { WebExtensionMessage } from '@dimensiondev/holoflows-kit'
 
 interface CashTagEvent {
     name: string
@@ -26,7 +26,7 @@ interface SwapDialogEvent {
     traderProps?: TraderProps
 }
 
-interface PluginTraderMessage {
+export interface TraderMessage {
     /**
      * View a cash tag
      */
@@ -51,8 +51,8 @@ interface PluginTraderMessage {
 }
 
 if (import.meta.webpackHot) import.meta.webpackHot.accept()
-export const PluginTraderMessages: WebExtensionMessage<PluginTraderMessage> =
-    createPluginMessage<PluginTraderMessage>(PLUGIN_IDENTIFIER)
+export const PluginTraderMessages: WebExtensionMessage<TraderMessage> =
+    createPluginMessage<TraderMessage>(PLUGIN_IDENTIFIER)
 export const PluginTraderRPC = createPluginRPC(
     PLUGIN_IDENTIFIER,
     () => import('./services'),
